+++ src/pages/tickets/centerTicket/createTicket.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import SubHeader from '@components/common/SubHeader';
+import SubHeader from '@components/common/SubHeader/SubHeader';
 import { LessonTypeEnum, TermUnitEnum } from '@/enums/Ticket';
 import Select from '@components/common/Select/Select';
 import Input from '@components/common/Input/Input';
